Make react and relay webpack externals
We are loading these as separate script tags on the page instead of
running them through webpack. This gives us a nice speed boost and makes
things a bit easier.

However, because we will be bringing in new components that require
these things explicitly, we want to make sure that we don't end up with
two copies of either. Setting the webpack externals configuration allows
us to use these as externals even in code that we do not control.