Add some pagination
To do this, I want it to have seamless pagination for the user by
automatically loading the next page on scroll.

I say that this is only "some" pagination, because it stops working
after a couple of pages due to my not fully wiring things up in the
database. This is because the `getIssues` function receives arguments
that include the `after` cursor and the `first` count which Relay pares
down to represent just what the next page worth it needs. I'm thinking
of a way to make this work better, but I haven't come up with anything
very good yet.

I was being rate limited while developing this, so I grabbed some API
keys from GitHub and added them here. These should be moved to
environment variables or something before productionizing.