Fix pagination
When getIssues is called, we get an `after` cursor and a `first` which
is the number of items to get after the `after` cursor. To figure out
how many things we actually need, we first need to find the item that
matches the cursor, and then add the value of `first` to that index plus
one. This will give us the count of things that we need to have loaded
which will tell us how many more we need to fetch at this time.

Relay won't issue new /graphql queries if the database returns the exact
amount of items it asked for or less, so we always need to fetch an
extra one if we can.