Add label colors
To pull this off, I needed to make the label schema more robust so that
I could actually query for the color in addition to the label's name.

The tricky part here that took me a while to figure out is that the
labels were all getting the same IDs, so they were changing to all be
the same as new pages of data loaded. This was happening because the
objects don't have an id key, so Relay's globalIdField didn't know what
to use. To fix this, I needed to provide an id resolver that resolved to
something unique on the objects, and URL seemed like the best candidate.