Add todo to schema
I was chatting in the #relay Slack channel and it seems like I don't
really benefit much from using this bit of code. If I have time, it
might not be a bad idea to come back and rewrite this so it doesn't need
it. In the interest of time, since things are actually working right
now, I'm going to leave it as is but I wanted to leave myself a note in
case I have time later to polish this up.