Add body to issue page
Issue bodies are in markdown format, so I brought in a markdown parser.
Before landing on marked, I tried remarkable [0] but that didn't work
out so well with webpack due to an issue in one of its dependencies
(Autolinker.js [1]). Next, I tried markdown-js [2] which seemed okay,
but it doesn't seem to have support for fenced code blocks or GitHub
flavored markdown [3], which I'd really like to have for this project
since I need to render markdown that comes from GitHub's API. marked
seems to support this better and it doesn't have problems with webpack.

[0]: https://github.com/jonschlinkert/remarkable
[1]: https://github.com/gregjacobs/Autolinker.js/issues/86
[2]: https://github.com/evilstreak/markdown-js
[3]: https://github.com/evilstreak/markdown-js/issues/41