Move issue id;number encoding into schema
I had originally tried this, but I believe that I was using `:` as my
delimiter at the time, which is used internally by graphql-relay, so it
didn't work and failed by resulting in the ID looking like it had
preferred the object's id property over my resolver. However, now that I
have a better understanding of how all of this stuff works, I can just
encode the ID and number in the schema and remove a bit of complexity.

The tradeoff here is that now I have the encoding and decoding spread
over two different files. I could solve this in a number of ways, such
as extracting the two functions into a module that schema and database
depend on, or by even keeping the two functions in the database and
exporting the encoding for the schema. However, I think this is fine for
now given that this is just a toy project and I don't intend on working
on it very much.